Get some types
