+++ components/about/Job.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
 
-import Job from "./job";
+import Job from "./Job";
 
 export const Template = (args) => <Job {...args} />;
